Added back blocklist read desc
I kept the formatting however some people might prefer /blocklist to /get block and it isn't bad to have other options. Of course it "makes the code longer", but consider that we are focusing on user experience and not just our code length, i.e. it makes it easier for our users to chat and isn't that the point?