Fixes https://github.com/FroggMaster/spooks-chatjs/issues/37
An extremely bad fix, but a fix nonetheless. **Totally usable and stable**. There is something wrong with backbone.js and it thinks a number string is a number or something, so you have to include another character i.e. '.' to make it THINK it's a string, then remove it during usage. Yep it's a very bad fix.